functions/requirements.txt
# Required for Cloud Functions framework
functions-framework==3.*

# Required for Firestore and Vertex AI Agent Engine integration
google-cloud-aiplatform[agent-engines]
google-cloud-firestore
google-cloud-aiplatform

firebase-admin
google-events

# Required for the Vertex AI Agent Development Kit
google-adk
